tests refactored with recommended changes implemented.

+++ cluster_device_monitor_test.py
@@ -6,22 +6,13 @@
 # information or intellectual property contained herein is strictly forbidden,
 # unless separate prior written permission has been obtained from Qumulo, Inc.
 
-# TODO list....
-# TODO: add tests for new QQ API exceptions changed in main
-# TODO: add tests for EmailMessage.send()
-# TODO: verify consistency of mock_send_email / mock_email
 
-
-# XXX Use something like this if your exception is going to call sys.exit but not as direct except:
-        # with self.assertRaisesRegex(SystemExit, 'ERROR'):
-        #     check_cluster_connectivity(CONFIG_DATA)
-
-
+import json
 import os
 import unittest
 
 from unittest import mock
-from qumulo.lib.request import RequestError # XXX remove? 
+from qumulo.lib.request import RequestError
 from typing import Any, Dict
 
 from cluster_device_monitor import (
@@ -88,7 +79,7 @@ class ParseConfigTest(unittest.TestCase):
 
     def test_bad_config_raises_error(self) -> None:
         with self.assertRaisesRegex(SystemExit, 'Configuration element missing.'):
-            _config = parse_config({'a': 'b'})
+            parse_config({'a': 'b'})
 
 
 @mock.patch(
@@ -97,24 +88,24 @@ class ParseConfigTest(unittest.TestCase):
 @mock.patch('cluster_device_monitor.socket')
 class SocketConnectivityTest(unittest.TestCase):
     def test_socket_connectivity(
-        self, _mock_socket: mock.MagicMock, mock_send_email: mock.MagicMock
+        self, _mock_socket: mock.MagicMock, mock_email: mock.MagicMock
     ) -> None:
         check_cluster_connectivity(CONFIG_DATA)
-        mock_send_email.assert_not_called()
+        mock_email.assert_not_called()
 
     def test_failed_socket_raises_connection_refused_error(
-        self, mock_socket: mock.MagicMock, mock_send_email: mock.MagicMock
+        self, mock_socket: mock.MagicMock, mock_email: mock.MagicMock
     ) -> None:
         mock_socket.socket.side_effect = ConnectionRefusedError()
         check_cluster_connectivity(CONFIG_DATA)
-        mock_send_email.assert_called_once()
+        mock_email.assert_called_once()
 
     def test_failed_socket_raises_timeout_error(
-        self, mock_socket: mock.MagicMock, mock_send_email: mock.MagicMock
+        self, mock_socket: mock.MagicMock, mock_email: mock.MagicMock
     ) -> None:
         mock_socket.socket.side_effect = TimeoutError()
         check_cluster_connectivity(CONFIG_DATA)
-        mock_send_email.assert_called_once()
+        mock_email.assert_called_once()
 
 
 class DeletePreviousStatusFileTest(unittest.TestCase):
@@ -131,23 +122,22 @@ class ClusterLoginTest(unittest.TestCase):
     def test_cluster_login(
         self, _mock_rest: mock.MagicMock, mock_email: mock.MagicMock
     ) -> None:
-        _return = cluster_login(CONFIG_DATA)
+        cluster_login(CONFIG_DATA)
         mock_email.assert_not_called()
 
     def test_cluster_failed_login_raises_timeout_error(
         self, mock_rest: mock.MagicMock, mock_email: mock.MagicMock
     ) -> None:
         mock_rest.side_effect = TimeoutError()
-        _response = cluster_login(CONFIG_DATA)
+        cluster_login(CONFIG_DATA)
         mock_email.assert_called_once()
 
-    # XXX FIX - BROKEN
-    # def test_cluster_bad_credentials_raises_request_error(
-    #     self, mock_rest: mock.MagicMock, mock_email: mock.MagicMock
-    # ) -> None:
-    #     mock_rest.side_effect = RequestError()
-    #     _response = cluster_login(CONFIG_DATA)
-    #     mock_email.assert_called_once()
+    def test_cluster_bad_credentials_raises_request_error(
+        self, mock_rest: mock.MagicMock, mock_email: mock.MagicMock
+    ) -> None:
+        mock_rest.side_effect = RequestError('401', 'Invalid Credentials')
+        cluster_login(CONFIG_DATA)
+        mock_email.assert_called_once()
 
 
 @mock.patch(
@@ -155,37 +145,40 @@ class ClusterLoginTest(unittest.TestCase):
 )
 @mock.patch('cluster_device_monitor.RestClient')
 class QQApiQueriesTest(unittest.TestCase):
-    def test_get_cluster_name(
+    def test_get_cluster_name_called(
         self, mock_rest: mock.MagicMock, mock_email: mock.MagicMock
     ) -> None:
-        _response = qq_api_query(mock_rest, CONFIG_DATA, 'cluster_name')
+        qq_api_query(mock_rest, CONFIG_DATA, 'cluster_name')
+        mock_rest.cluster.get_cluster_conf.assert_called_once()
         mock_email.assert_not_called()
 
-    def test_get_cluster_version(
+    def test_get_cluster_version_called(
         self, mock_rest: mock.MagicMock, mock_email: mock.MagicMock
     ) -> None:
-        _response = qq_api_query(mock_rest, CONFIG_DATA, 'qq_version')
+        qq_api_query(mock_rest, CONFIG_DATA, 'qq_version')
+        mock_rest.version.version.assert_called_once()
         mock_email.assert_not_called()
 
-    def test_get_cluster_time(
+    def test_get_cluster_time_called(
         self, mock_rest: mock.MagicMock, mock_email: mock.MagicMock
     ) -> None:
-        _response = qq_api_query(mock_rest, CONFIG_DATA, 'cluster_time')
+        qq_api_query(mock_rest, CONFIG_DATA, 'cluster_time')
+        mock_rest.time_config.get_time_status.assert_called_once()
         mock_email.assert_not_called()
 
-    def test_get_cluster_uuid(
+    def test_get_cluster_uuid_called(
         self, mock_rest: mock.MagicMock, mock_email: mock.MagicMock
     ) -> None:
-        _response = qq_api_query(mock_rest, CONFIG_DATA, 'cluster_uuid')
+        qq_api_query(mock_rest, CONFIG_DATA, 'cluster_uuid')
+        mock_rest.node_state.get_node_state.assert_called_once()
         mock_email.assert_not_called()
 
-    # XXX FIX - BROKEN
     def test_api_timeout_raises_timeout_error(
         self, mock_rest: mock.MagicMock, mock_email: mock.MagicMock
     ) -> None:
         mock_rest.side_effect = TimeoutError()
-        _response = cluster_login(CONFIG_DATA)
-        mock_email.assert_called_once()    
+        cluster_login(CONFIG_DATA)
+        mock_email.assert_called_once()
 
 
 @mock.patch(
@@ -193,93 +186,41 @@ class QQApiQueriesTest(unittest.TestCase):
 )
 @mock.patch('cluster_device_monitor.RestClient')
 class RetrieveStatusOfClusterDevicesTest(unittest.TestCase):
-    def setUp(self) -> None:
-        self.sample_node_real_dict = {
-            'id': 1,
-            'node_status': 'online',
-            'node_name': 'CoffeeTime-1',
-            'model_number': 'QVIRT',
-            'serial_number': 'QVIRT',
-        }
-        self.sample_drive_real_dict = {
-            'id': '1.1',
-            'node_id': 1,
-            'slot': 1,
-            'state': 'healthy',
-            'disk_type': 'SSD',
-            'disk_model': 'Virtual_disk',
-            'disk_serial_number': '',
-            'capacity': '10467934208',
-        }
-
-    def test_retrieve_status_of_nodes(
+    def test_retrieve_status_of_nodes_called(
         self, mock_rest: mock.MagicMock, mock_email: mock.MagicMock
     ) -> None:
-        expected = {
-            'id': 1,
-            'node_status': 'online',
-            'node_name': 'CoffeeTime-1',
-            'model_number': 'QVIRT',
-            'serial_number': 'QVIRT',
-        }
-        mock_rest.cluster.list_nodes.return_value = [self.sample_node_real_dict]
-        status_of_nodes = retrieve_status_of_cluster_devices(
+        retrieve_status_of_cluster_devices(
             mock_rest, CONFIG_DATA, 'nodes'
         )
-        self.assertEqual(status_of_nodes['nodes'][0], expected)
-        expected['node_status'] = 'offline'
-        status_of_nodes['nodes'][0]['node_status'] = 'offline'
-        self.assertEqual(status_of_nodes['nodes'][0], expected)
+        mock_rest.cluster.list_nodes.called_once()
         mock_email.assert_not_called()
 
-    def test_retrieve_status_of_drives(
+    def test_retrieve_status_of_drives_called(
         self, mock_rest: mock.MagicMock, mock_email: mock.MagicMock
     ) -> None:
-        expected = {
-            'id': '1.1',
-            'node_id': 1,
-            'slot': 1,
-            'state': 'healthy',
-            'disk_type': 'SSD',
-            'disk_model': 'Virtual_disk',
-            'disk_serial_number': '',
-            'capacity': '10467934208',
-        }
-        mock_rest.cluster.get_cluster_slots_status.return_value = [
-            self.sample_drive_real_dict
-        ]
-        status_of_drives = retrieve_status_of_cluster_devices(
+        retrieve_status_of_cluster_devices(
             mock_rest, CONFIG_DATA, 'drives'
         )
-        self.assertEqual(status_of_drives['drives'][0], expected)
-        expected['state'] = 'unhealthy'
-        status_of_drives['state'] = 'unhealthy'
-        status_of_drives['drives'][0]['state'] = 'unhealthy'
-        self.assertEqual(status_of_drives['drives'][0], expected)
+        mock_rest.cluster.get_cluster_slots_status.called_once()
         mock_email.assert_not_called()
 
-    # XXX - BROKEN
-    # def test_api_timeout_raises_error(
-    #     self, mock_rest: mock.MagicMock, mock_email: mock.MagicMock
-    # ) -> None:
-    #     mock_rest.cluster.list_nodes.side_effect = Exception()
-    #     retrieve_status_of_cluster_devices(mock_rest, CONFIG_DATA, 'nodes')
-    #     mock_email.assert_called_once()
+    def test_api_timeout_raises_error(
+        self, mock_rest: mock.MagicMock, mock_email: mock.MagicMock
+    ) -> None:
+        mock_rest.cluster.list_nodes.side_effect = TimeoutError()
+        retrieve_status_of_cluster_devices(mock_rest, CONFIG_DATA, 'nodes')
+        mock_email.assert_called_once()
 
 
 class PreserveClusterStatusTest(unittest.TestCase):
     def setUp(self) -> None:
-        self.cluster_status = {
+        self.test_cluster_status = {
             'nodes': [
                 {
                     'id': 1,
                     'node_status': 'online',
-                    'node_name': 'CoffeeTime-1',
-                    'uuid': '10a1c7aa-fb99-48a1-8dc3-b34b96777742',
-                    'label': '00:50:56:bf:68:82',
                     'model_number': 'QVIRT',
                     'serial_number': 'QVIRT',
-                    'mac_address': '00:50:56:bf:68:82',
                 }
             ],
             'drives': [
@@ -293,34 +234,32 @@ class PreserveClusterStatusTest(unittest.TestCase):
                     'disk_model': 'Virtual_disk',
                     'disk_serial_number': '1234',
                     'capacity': '10467934208',
-                    'raw_capacity': '10737418240',
-                    'minimum_raw_capacity': '10737418240',
-                    'high_endurance': False,
-                    'drive_bay': '',
-                    'led_pattern': 'LED_PATTERN_NORMAL',
                 }
             ],
         }
 
     def test_cluster_status_file_created_and_populated(self) -> None:
-        preserve_cluster_status(self.cluster_status)
+        preserve_cluster_status(self.test_cluster_status)
         self.assertIn('cluster_status.json', os.listdir())
-        self.assertEqual(self.cluster_status['nodes'][0]['id'], 1)
-        self.assertEqual(self.cluster_status['nodes'][0]['node_status'], 'online')
-        self.assertEqual(self.cluster_status['nodes'][0]['model_number'], 'QVIRT')
-        self.assertEqual(self.cluster_status['nodes'][0]['serial_number'], 'QVIRT')
-        self.assertEqual(self.cluster_status['drives'][0]['id'], '1.1')
-        self.assertEqual(self.cluster_status['drives'][0]['node_id'], 1)
-        self.assertEqual(self.cluster_status['drives'][0]['slot'], 1)
-        self.assertEqual(self.cluster_status['drives'][0]['state'], 'healthy')
-        self.assertEqual(self.cluster_status['drives'][0]['disk_type'], 'SSD')
-        self.assertEqual(self.cluster_status['drives'][0]['disk_model'], 'Virtual_disk')
-        self.assertEqual(self.cluster_status['drives'][0]['disk_serial_number'], '1234')
-        self.assertEqual(self.cluster_status['drives'][0]['capacity'], '10467934208')
+
+        with open('cluster_status.json', 'r') as file:
+            json_file = json.load(file)
+            self.assertEqual(json_file['nodes'][0]['id'], 1)
+            self.assertEqual(json_file['nodes'][0]['node_status'], 'online')
+            self.assertEqual(json_file['nodes'][0]['model_number'], 'QVIRT')
+            self.assertEqual(json_file['nodes'][0]['serial_number'], 'QVIRT')
+            self.assertEqual(json_file['drives'][0]['id'], '1.1')
+            self.assertEqual(json_file['drives'][0]['node_id'], 1)
+            self.assertEqual(json_file['drives'][0]['slot'], 1)
+            self.assertEqual(json_file['drives'][0]['state'], 'healthy')
+            self.assertEqual(json_file['drives'][0]['disk_type'], 'SSD')
+            self.assertEqual(json_file['drives'][0]['disk_model'], 'Virtual_disk')
+            self.assertEqual(json_file['drives'][0]['disk_serial_number'], '1234')
+            self.assertEqual(json_file['drives'][0]['capacity'], '10467934208')
 
     def tearDown(self) -> None:
-        os.remove('cluster_status.json')
-        self.assertTrue('cluster_status.json' not in os.listdir())
+        if os.path.exists('cluster_status.json'):
+            os.remove('cluster_status.json')
 
 
 class CheckForUnhealthyObjectsTest(unittest.TestCase):
@@ -370,36 +309,27 @@ class BuildEmailTest(unittest.TestCase):
             'Event 1': {
                 'id': 2,
                 'node_status': 'offline',
-                'node_name': 'CoffeeTime-2',
-                'uuid': 'cbdea0e3-1659-48af-b15b-e97dbbeefd04',
-                'label': '00:50:56:bf:f1:57',
                 'model_number': 'QVIRT',
                 'serial_number': '',
-                'mac_address': '',
             }
         }
-        self.qq_version = '4.1.0'
-        self.cluster_name = 'CoffeeTime'
-        self.cluster_uuid = 'cf83e828-7ef7-4368-a75b-3b972d10f2c6'  # len() = 36
-        self.cluster_time = '2021-07-23T17:35:29.189487278Z UTC'  # len() = 34
 
     def test_qq_api_query_success(
         self, mock_rest: mock.MagicMock, mock_email: mock.MagicMock
     ) -> None:
-        _email_alert = populate_alert_email_body(
+        populate_alert_email_body(
             self.alert_data, mock_rest, CONFIG_DATA
         )
         mock_email.assert_not_called()
 
-    # XXX FIX - BROKEN
-    # def test_qq_api_query_timeout_raises_error(
-    #     self, mock_rest: mock.MagicMock, mock_email: mock.MagicMock
-    # ) -> None:
-    #     mock_rest.cluster.get_cluster_conf.side_effect = Exception()
-    #     _email_alert = populate_alert_email_body(
-    #         self.alert_data, mock_rest, CONFIG_DATA
-    #     )
-    #     mock_email.assert_called_once()
+    def test_api_timeout_raises_error(
+        self, mock_rest: mock.MagicMock, mock_email: mock.MagicMock
+    ) -> None:
+        mock_rest.cluster.get_cluster_conf.side_effect = TimeoutError()
+        populate_alert_email_body(
+            self.alert_data, mock_rest, CONFIG_DATA
+        )
+        mock_email.assert_called_once()
 
     def test_email_contains_alert_data(
         self, mock_rest: mock.MagicMock, _mock_email: mock.MagicMock
@@ -411,47 +341,36 @@ class BuildEmailTest(unittest.TestCase):
         self.assertIn(self.alert_data['Event 1']['model_number'], email_alert)
 
 
-# XXX this will need fixed as send_email doesn't exist anymore
-# @mock.patch('cluster_device_monitor.smtplib')
-# class SendEmailTest(unittest.TestCase):
-#     def test_can_send_email(self, _mock_smtp: mock.MagicMock) -> None:
-#         send_email(EML)
+@mock.patch('cluster_device_monitor.EmailMessage.send')
+class GenerateEventAlertEmailTest(unittest.TestCase):
+    def test_send_email_success(self, mock_email: mock.MagicMock) -> None:
+        generate_event_alert_email(CONFIG_DATA, 'foo')
+        mock_email.assert_called_once()
+
+    def test_send_email_failure_raises_error(
+        self, mock_email: mock.MagicMock
+    ) -> None:
+        mock_email.side_effect = Exception()
+        with self.assertRaisesRegex(SystemExit, 'Check connection to SMTP server.'):
+            generate_event_alert_email(CONFIG_DATA, 'foo')
 
 
-# # XXX this will need fixed as send_email doesn't exist anymore
-# @mock.patch('cluster_device_monitor.send_email')
-# class GenerateEventAlertEmailTest(unittest.TestCase):
-#     def test_send_email_success(self, mock_send_email: mock.MagicMock) -> None:
-#         generate_event_alert_email(CONFIG_DATA, 'foo')
-#         mock_send_email.assert_called_once()
-
-#     def test_send_email_failure_raises_error(
-#         self, mock_send_email: mock.MagicMock
-#     ) -> None:
-#         mock_send_email.side_effect = Exception()
-#         with self.assertRaisesRegex(SystemExit, 'Check connection to SMTP server.'):
-#             generate_event_alert_email(CONFIG_DATA, 'foo')
-
-
-# XXX this needs to be changed to reflect new function name.. problem, not timeout
-@mock.patch('cluster_device_monitor.send_email')
-class GenerateTimeoutAlertEmailTest(unittest.TestCase):
+@mock.patch('cluster_device_monitor.EmailMessage.send')
+class GenerateScriptProblemEmailTest(unittest.TestCase):
     def setUp(self) -> None:
         self.error = 'Error 403: Access Denied.'
 
-    # XXX FIX - BROKEN
-    # def test_send_email_success(self, mock_send_email: mock.MagicMock) -> None:
-    #     with self.assertRaisesRegex(SystemExit, 'EMAIL SENT'):
-    #         generate_script_problem_email(self.error, CONFIG_DATA)
-    #     mock_send_email.assert_called_once()
+    def test_send_email_success(self, mock_email: mock.MagicMock) -> None:
+        with self.assertRaisesRegex(SystemExit, 'EMAIL SENT'):
+            generate_script_problem_email(self.error, CONFIG_DATA)
+        mock_email.assert_called_once()
 
-    # XXX FIX - BROKEN
-    # def test_send_email_failure_raises_error(
-    #     self, mock_send_email: mock.MagicMock
-    # ) -> None:
-    #     mock_send_email.side_effect = Exception()
-    #     with self.assertRaisesRegex(SystemExit, 'Unable to send email.'):
-    #         generate_script_problem_email(self.error, CONFIG_DATA)
+    def test_send_email_failure_raises_error(
+        self, mock_email: mock.MagicMock
+    ) -> None:
+        mock_email.side_effect = Exception()
+        with self.assertRaisesRegex(SystemExit, 'Unable to send email.'):
+            generate_script_problem_email(self.error, CONFIG_DATA)
 
 
 if __name__ == '__main__':
